requirements.txt
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.1.8
colorama==0.4.6
fastapi==0.135.1
greenlet==3.4.0
gTTS==2.5.4
h11==0.16.0
httptools==0.7.1
idna==3.11
Mako==1.3.11
MarkupSafe==3.0.3
psycopg2==2.9.11
psycopg2-binary==2.9.11
pydantic==2.13.2
pydantic-settings==2.13.1
pydantic_core==2.46.2
PyMuPDF==1.27.2.2
python-dotenv==1.2.2
python-multipart==0.0.26
PyYAML==6.0.3
requests==2.33.1
SQLAlchemy==2.0.48
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.44.0
watchfiles==1.1.1
websockets==16.0
